refactor: centralize disposition writing into result_collector
Move all set_disposition calls out of pipeline.py, initial_pipeline.py,
and processing_recovery.py into result_collector.py.

- Add write_node_level_disposition() helper for node-level skip/passthrough,
  replacing copy-pasted blocks in pipeline.py and initial_pipeline.py
- Move write_record_dispositions() from processing_recovery.py to
  result_collector.py, replacing direct set_disposition calls with the
  shared _safe_set_disposition wrapper
- set_disposition calls now only exist in result_collector.py (centralized)
  and executor.py (executor-level failure/skip)
- Net -19 lines across 9 files
