+++ .changes/unreleased/Under the Hood-20260428-304000.yaml
@@ -1,0 +1,3 @@
+kind: Under the Hood
+body: "Centralize disposition writing: node-level and batch record dispositions now route through result_collector.py instead of being scattered across pipeline, initial_pipeline, and processing_recovery modules"
+time: 2026-04-28T30:40:00.000000Z
